InputContainer와 ListContainer 생성
