+++ src/App.test.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
+import { useSelector } from 'react-redux';
+import { render } from '@testing-library/react';
 
-import { render, fireEvent } from '@testing-library/react';
-
-import { useSelector, useDispatch } from 'react-redux';
 import App from './App';
 
 jest.mock('react-redux');
@@ -15,40 +14,15 @@ beforeEach(() => {
 
   useSelector.mockImplementation((selector) => selector({
     tasks,
-    taskTitle: '',
   }));
 });
 
-describe('App', () => {
-  const dispatch = jest.fn();
+test('App', () => {
+  const { getByText } = render(<App />);
 
-  useDispatch.mockImplementation(() => dispatch);
-
-  it('renders tasks', () => {
-    const { getByText } = render((
-      <App />
-    ));
-
-    expect(getByText('Task-1')).not.toBeNull();
-    expect(getByText('Task-2')).not.toBeNull();
-  });
-
-  it('listens change event', () => {
-    const { getByLabelText } = render((
-      <App />
-    ));
-
-    fireEvent.change(getByLabelText('할 일'), {
-      target: { value: '무언가 하기' },
-    });
-
-    expect(dispatch).toBeCalledWith({
-      type: 'updateTaskTitle',
-      payload: {
-        taskTitle: '무언가 하기',
-      },
-    });
-  });
+  expect(getByText('추가')).not.toBeNull();
+  expect(getByText('Task-1')).not.toBeNull();
+  expect(getByText('Task-2')).not.toBeNull();
 
   // TODO: 통합 테스트 코드 작성
   // CodeceptJS => 실제 브라우저에서 사용자 테스트 실행 가능.
