+++ src/InputContainer.jsx
@@ -1,0 +1,30 @@
+import React from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+
+import { addTask, updateTaskTitle } from './action';
+
+import Input from './Input';
+
+export default function InputContainer() {
+  const dispatch = useDispatch();
+
+  const { taskTitle } = useSelector((state) => ({
+    taskTitle: state.taskTitle,
+  }));
+
+  function handleChangeTitle(event) {
+    dispatch(updateTaskTitle(event.target.value));
+  }
+
+  function handleClickAddTask() {
+    dispatch(addTask());
+  }
+
+  return (
+    <Input
+      value={taskTitle}
+      onChange={handleChangeTitle}
+      onClick={handleClickAddTask}
+    />
+  );
+}
